requirements.txt
# MySQL schema export, Text-to-SQL (ModelScope OpenAI-compatible), and SQL validation
PyMySQL==1.1.1
sqlglot==23.16.0
openai>=1.30.0
python-dotenv==1.0.1
tabulate==0.9.0
sentence-transformers>=2.6.1
faiss-cpu>=1.7.4
